+++ services/lawyer-profile-service/src/main/java/com/project/lawyerprofileservice/dto/LawyerResponseDTO.java
@@ -4,6 +4,7 @@ import com.project.lawyerprofileservice.model.Specialization;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -24,7 +25,7 @@ public class LawyerResponseDTO {
 
     private String bio;
 
-    private Double consultationFee;
+    private BigDecimal consultationFee;
 
     private String location;
